Move `Map` out into its own file
