+++ docs/GenerateMarkdown.go
@@ -10,15 +10,6 @@ import (
 	md "github.com/Shresht7/Scribe/markdown"
 )
 
-// Map applies the given function to each item in the given slice and returns a new slice with the results.
-func Map[T, R any](items []T, fn func(T) R) []R {
-	result := make([]R, len(items))
-	for i, item := range items {
-		result[i] = fn(item)
-	}
-	return result
-}
-
 // GenerateMarkdown generates the markdown for the given command.
 func GenerateMarkdown(cmd *cobra.Command) (string, error) {
